refactor: restructure project and fix CI (#15)
Rust:
- Extract shared helpers (get_groups, df_to_hashmap, multivariate variants)
  into src/utils.rs, removing duplication across 9 distance files
- Delete dead src/dtw_extended.rs (not referenced in lib.rs)
- Clean up commented-out code blocks in lib.rs

Python:
- Populate decomposition/__init__.py and models/__init__.py with re-exports
- Update polars_ts/__init__.py to expose subpackage symbols

Tests:
- Extract shared fixtures into tests/conftest.py
- Organize tests into tests/distance/, tests/decomposition/, tests/metrics/
- Move _to_dict helper to tests/distance/conftest.py

CI:
- Upgrade deprecated GitHub Actions (v3 → v4/v5)

All 114 tests pass with no regressions.
