+++ polars_ts/decomposition/__init__.py
@@ -1,0 +1,5 @@
+from polars_ts.decomposition.fourier_decomposition import fourier_decomposition
+from polars_ts.decomposition.seasonal_decomposition import seasonal_decomposition
+from polars_ts.decomposition.seasonal_decompose_features import seasonal_decompose_features
+
+__all__ = ["fourier_decomposition", "seasonal_decomposition", "seasonal_decompose_features"]
