+++ polars_ts/models/__init__.py
@@ -1,0 +1,3 @@
+from polars_ts.models.scum import SCUM
+
+__all__ = ["SCUM"]
