+++ src/ddtw.rs
@@ -1,23 +1,11 @@
 use polars::prelude::*;
-use std::collections::HashMap;
 use std::sync::Arc;
 use pyo3::prelude::*;
 use pyo3_polars::PyDataFrame;
 use pyo3::PyResult;
 use rayon::prelude::*;
 
-/// Groups a DataFrame by "unique_id" and aggregates the "y" column.
-/// (Casting "unique_id" as Utf8 and "y" as Float64.)
-fn get_groups(df: &DataFrame) -> Result<LazyFrame, PolarsError> {
-    Ok(df.clone().lazy()
-        .select([
-            col("unique_id").cast(DataType::String),
-            col("y").cast(DataType::Float64)
-        ])
-        .group_by([col("unique_id")])
-        .agg([col("y")])
-    )
-}
+use crate::utils::{get_groups, df_to_hashmap};
 
 /// Compute the derivative of a time series using the method from Keogh & Pazzani (2001).
 ///
@@ -39,9 +27,6 @@ fn compute_derivative(q: &[f64]) -> Vec<f64> {
         let term1 = q[i] - q[i-1];
         let term2 = (q[i+1] - q[i-1]) / 2.0;
         let derivative_i = (term1 + term2) / 2.0;
-
-        // Alternatively, simplified as in the Python code:
-        // let derivative_i = 0.25 * q[i+1] + 0.5 * q[i] - 0.75 * q[i-1];
 
         derivative.push(derivative_i);
     }
@@ -91,46 +76,6 @@ fn ddtw_distance(a: &[f64], b: &[f64]) -> f64 {
 
     // Compute DTW on derivatives
     dtw_distance(&a_derivative, &b_derivative)
-}
-
-/// Optimized conversion of a grouped DataFrame into a HashMap mapping id -> Vec<f64>.
-fn df_to_hashmap(df: &DataFrame) -> HashMap<String, Vec<f64>> {
-    // Retrieve the columns.
-    let unique_id_col = df.column("unique_id").expect("expected column unique_id");
-    let y_col = df.column("y").expect("expected column y");
-
-    // Collect unique IDs into a Vec<String>.
-    let unique_ids: Vec<String> = unique_id_col
-        .str()
-        .expect("expected utf8 column for unique_id")
-        .into_no_null_iter()
-        .map(|s| s.to_string())
-        .collect();
-
-    // Collect each list element into a Vec<f64>.
-    let y_lists: Vec<Vec<f64>> = y_col
-        .list()
-        .expect("expected a List type for y")
-        .into_iter()
-        .map(|opt_series| {
-            let series = opt_series.expect("null entry in 'y' list column");
-            series
-                .f64()
-                .expect("expected a f64 Series inside the list")
-                .into_no_null_iter()
-                .collect::<Vec<f64>>()
-        })
-        .collect();
-
-    // Sanity-check that we have the same number of ids and y vectors.
-    assert_eq!(unique_ids.len(), y_lists.len(), "Mismatched lengths in unique_ids and y_lists");
-
-    // Build the HashMap in parallel.
-    let hashmap: HashMap<String, Vec<f64>> = (0..unique_ids.len())
-        .into_par_iter()
-        .map(|i| (unique_ids[i].clone(), y_lists[i].clone()))
-        .collect();
-    hashmap
 }
 
 /// Compute pairwise DDTW distances between time series in two DataFrames,
