+++ src/dtw.rs
@@ -1,22 +1,11 @@
 use polars::prelude::*;
-use std::collections::HashMap;
+use std::sync::Arc;
 use pyo3::prelude::*;
 use pyo3_polars::PyDataFrame;
 use pyo3::PyResult;
 use rayon::prelude::*;
 
-/// Groups a DataFrame by "unique_id" and aggregates the "y" column.
-/// (Casting "unique_id" as Utf8 and "y" as Float64.)
-fn get_groups(df: &DataFrame) -> Result<LazyFrame, PolarsError> {
-    Ok(df.clone().lazy()
-        .select([
-            col("unique_id").cast(DataType::String),
-            col("y").cast(DataType::Float64)
-        ])
-        .group_by([col("unique_id")])
-        .agg([col("y")])
-    )
-}
+use crate::utils::{get_groups, df_to_hashmap};
 
 // ---------------------------------------------------------------------------
 // DTW distance functions
@@ -87,10 +76,6 @@ fn dtw_itakura(a: &[f64], b: &[f64], max_slope: f64) -> f64 {
         curr[0] = f64::MAX;
         for j in 1..=m {
             // Check Itakura parallelogram bounds.
-            // The parallelogram constrains j/i and (m-j)/(n-i) by max_slope
-            // and 1/max_slope. Equivalently, for position (i, j):
-            //   j >= i / max_slope  AND  j <= i * max_slope
-            //   j >= m - (n - i) * max_slope  AND  j <= m - (n - i) / max_slope
             let fi = i as f64;
             let fj = j as f64;
             let lower1 = fi / max_slope;
@@ -320,50 +305,6 @@ fn compute_dtw(a: &[f64], b: &[f64], method: &str, param: f64) -> f64 {
         "fast" => fast_dtw(a, b, param as usize),
         _ => dtw_distance(a, b),
     }
-}
-
-/// Optimized conversion of a grouped DataFrame into a HashMap mapping id -> Vec<f64>.
-///
-/// This version first collects the "unique_id" column and the list-of-f64
-/// from the "y" column into two vectors. Then, using a parallel index loop,
-/// it zips them together into a HashMap.
-fn df_to_hashmap(df: &DataFrame) -> HashMap<String, Vec<f64>> {
-    // Retrieve the columns.
-    let unique_id_col = df.column("unique_id").expect("expected column unique_id");
-    let y_col = df.column("y").expect("expected column y");
-
-    // Collect unique IDs into a Vec<String>.
-    let unique_ids: Vec<String> = unique_id_col
-        .str()
-        .expect("expected utf8 column for unique_id")
-        .into_no_null_iter()
-        .map(|s| s.to_string())
-        .collect();
-
-    // Collect each list element into a Vec<f64>.
-    let y_lists: Vec<Vec<f64>> = y_col
-        .list()
-        .expect("expected a List type for y")
-        .into_iter()
-        .map(|opt_series| {
-            let series = opt_series.expect("null entry in 'y' list column");
-            series
-                .f64()
-                .expect("expected a f64 Series inside the list")
-                .into_no_null_iter()
-                .collect::<Vec<f64>>()
-        })
-        .collect();
-
-    // Sanity-check that we have the same number of ids and y vectors.
-    assert_eq!(unique_ids.len(), y_lists.len(), "Mismatched lengths in unique_ids and y_lists");
-
-    // Build the HashMap in parallel.
-    let hashmap: HashMap<String, Vec<f64>> = (0..unique_ids.len())
-        .into_par_iter()
-        .map(|i| (unique_ids[i].clone(), y_lists[i].clone()))
-        .collect();
-    hashmap
 }
 
 /// Compute pairwise DTW distances between time series in two DataFrames,
