+++ tests/decomposition/test_fourier_decomposition.py
