+++ tests/decomposition/test_seasonal_decomposition.py
