+++ tests/distance/test_dtw.py
@@ -2,79 +2,7 @@ import polars as pl
 import pytest
 from polars_ts_rs.polars_ts_rs import compute_pairwise_dtw
 
-
-# ---------------------------------------------------------------------------
-# Fixtures
-# ---------------------------------------------------------------------------
-
-@pytest.fixture
-def two_series():
-    """Two simple time series A and B that differ by one point."""
-    return pl.DataFrame({
-        "unique_id": ["A"] * 4 + ["B"] * 4,
-        "y": [1.0, 2.0, 3.0, 4.0, 1.0, 2.0, 3.0, 5.0],
-    })
-
-
-@pytest.fixture
-def three_series():
-    """Three time series: A ascending, B similar to A, C reversed."""
-    return pl.DataFrame({
-        "unique_id": ["A"] * 4 + ["B"] * 4 + ["C"] * 4,
-        "y": [1.0, 2.0, 3.0, 4.0, 1.0, 2.0, 3.0, 5.0, 4.0, 3.0, 2.0, 1.0],
-    })
-
-
-@pytest.fixture
-def identical_series():
-    """Two identical time series."""
-    return pl.DataFrame({
-        "unique_id": ["A"] * 4 + ["B"] * 4,
-        "y": [1.0, 2.0, 3.0, 4.0, 1.0, 2.0, 3.0, 4.0],
-    })
-
-
-@pytest.fixture
-def shifted_series():
-    """Two series where one is a shifted version — constraints matter here."""
-    return pl.DataFrame({
-        "unique_id": ["A"] * 8 + ["B"] * 8,
-        "y": [1.0, 1.0, 1.0, 1.0, 5.0, 5.0, 5.0, 5.0,
-              5.0, 5.0, 5.0, 5.0, 1.0, 1.0, 1.0, 1.0],
-    })
-
-
-@pytest.fixture
-def single_series():
-    """A single time series — no pairs to compare."""
-    return pl.DataFrame({
-        "unique_id": ["A"] * 4,
-        "y": [1.0, 2.0, 3.0, 4.0],
-    })
-
-
-@pytest.fixture
-def int_id_series():
-    """Time series with integer unique_id."""
-    return pl.DataFrame({
-        "unique_id": [1] * 4 + [2] * 4,
-        "y": [1.0, 2.0, 3.0, 4.0, 4.0, 3.0, 2.0, 1.0],
-    })
-
-
-# ---------------------------------------------------------------------------
-# Helpers
-# ---------------------------------------------------------------------------
-
-def _to_dict(df: pl.DataFrame) -> dict:
-    """Convert pairwise result to {(id1, id2): distance} dict, sorted keys."""
-    rows = df.to_dicts()
-    dist_col = [c for c in df.columns if c not in ("id_1", "id_2")][0]
-    result = {}
-    for r in rows:
-        key = tuple(sorted([str(r["id_1"]), str(r["id_2"])]))
-        result[key] = r[dist_col]
-    return result
+from tests.distance.conftest import _to_dict
 
 
 ALL_METHODS = [
